Uploading first sketch of 3D-EnVAR
Started modifications of 3D-VAR scripts for ensemble integration,
next stage is to write Rocoto metatasks for ensemble loop and test
initial analysis from cold start.  Once cold start is validated,
need to integrate and test the observer version of GSI for EnKF
and test analysis of ensemble perturbations by EnKF.  If this works,
finally can loop the outputs of the GSI/EnKF analyses into the WRFDA
steps for boundary conditions and full cycling.