Further integration of enkf, testing new cycling dependency
New cycle with free forecast didn't start automatically with the
metatask dependence, and was loaded inefficiently in the first place
using such a tag.  This is replaced with a wrf ensemble taks dependency
that needs to be tested. Currently running sensitivity test to quilting
where no quilting appears to improve performance with the smaller high
res domain configuration.  EnKF integration needs dynamic loading of
grid variables similar to GSI script, need testing of reading
with ncdump or other utility.