Re-write of cycle to start with WRF forecast / dual res for control and ens
First sketch at re-writing to start with WRF forecast removing the
issue with the background data, where we can utilize the same zero hour
initialization and forecast zero hour for the background boundary conditions,
smoothed with WRFDA to match the cyling analysis (or free forecast at time 0).
Restructured to switch the ensemble and control domains automatically,
running for now 3D-VAR alone on d02.