Uploading GSI diagnostic plot, note GSI bug
The number of observations monitored doesn't remain fixed in v3.7
over the loops so that one cannot accurately compute the validation
error for the analysis versus forecast RMSE on an independent set
of observations, but it doesn't seem like there is a good
alternative and the number of monitored observations doesn't change
too much in some initial scanning.  Will try a 3D version shortly
for further diagnostics.