EnKF Control flow working through first forecast, staging test of downscale
Bug fix in the refactor of the EnKF based control flow, this appears to
work but awaiting tests of a DA cycle to confirm.  Downscaling experiment
is now staged as a variation on the EnKF and the WPS versions, where
we leave out the perturbation cycle, but include the ensemble forecast
in WRF.  Will start developing GFS AWS utility for possible easier
forcing data transfer, then test additional satellite obs sources.