Major revision and cleanup of gsi.sh, minor bug fixes with missing fix files
New path structure is intended to smooth the execution of gsi across
multiple machines and to make the workflow based around experiment configurations
including the fix files and an additional auxilliary satellite list file so
that options are not hard coded outside of an experiment's local definitions.
This is intended to allow multiple simulation configurations to run on
the same clone simultaneously, thus to scale.