Merged gridstat codes and GSI codes for analyzing data, merging Mode codes and static ensembles
Starting to work on further efficiency with control flow in order to rerun tuning experiments
with static perts in 3D-EnVAR when not cycling ensemble perturbations.  These can be reused
in order to re-run the control alone at a lower cost, allowing for rapid tuning of beta parameters
localiation parameters and filtering obs.  Pending proof of concept with new structure
tomorrow.