Testing RAP background / fix, bug fixes with grep of status
Testing an implementation of the RAP background error and fix files,
generalized the code to implement a linking strategy for generic names
of the GSI fix files.  Fixed bug in terms of searching for cycle with utility
script, now is generalized to trigger cycles for both normal cycling and
extended forecasts.  Will determine if this will run automatically or if
still needs booting after the initiation of the restart run for the extended
forecast.  Will move to Cylc conversion ASAP due to general Rocoto issues.