Successful restart run from old extended forecast files
Fixed missing comment for GSI to stop verbose logs.  WRF driver script is now
edited to handle restart runs for extended forecasts drawn from cycling conditions.
Needed to add step to this condition to re-link the boundary conditions
for the full length of a new forecast horizon.  Keeping template for
how to run this, preparing for cleanup, merger and cylc rewrite.