Testing bugfix for restart runs, cleaning archive material from namelists
Needed to set a parameter to make restart run restart writeout timer to be
read from namelist.  This is now implemented to possibly bugfix issue with
current restart runs in which they still use the 6-hour restart interval from
the original run, mistakenly instead of using the forecast horizon that is written
to the namelist.  Removed old archive materials from the namelists to make these
more static as templates that are easy for reading and filling.