Testing new approach to hack rocoto
Building in dummy task within cycle that will trigger based on the completion of
the wrf short-range, six-hour cycle and grep the logs for the first task to boot in
the next cycle.  Inter-cycle dependencies have been fully removed, where the dummy
task will be used to force the next cycle to trigger when Rocoto fails.